Fix NuGet clean script

Related work items: #187301

+++ Src/Scripts/make-free-space-vnext.ps1
@@ -17,17 +17,17 @@
 # Updated the build agent cleaner to be less cody and use Get-child recursion with remove-item.
 #
 param(
-	[Parameter(Mandatory=$false)][string]$drive,
-	[Parameter(Mandatory=$false)][ValidateSet("agent", "nuget")][String]$strategy,
-	[Parameter(Mandatory=$false)][Int]$olderThan = 24
+    [Parameter(Mandatory=$false)][string]$drive,
+    [Parameter(Mandatory=$false)][ValidateSet("agent", "nuget")][string]$strategy,
+    [Parameter(Mandatory=$false)][Int]$olderThan = 24
 )
 
 if ([string]::IsNullOrWhiteSpace($drive)) {
-	$drive = "C"
+    $drive = "C"
 }
 
 if ([string]::IsNullOrWhiteSpace($strategy)) {
-	$strategy = "agent"
+    $strategy = "agent"
 }
 
 # If disk space is less than $percentageAtWhichToPanic, no folders younger than this will be deleted.
@@ -43,36 +43,38 @@ if ([string]::IsNullOrWhiteSpace($strategy)) {
 # Get the percentage space used for a drive
 Function Get-PercentUsed($drive) {
     $s = "DeviceID='" + $drive + ":'"
-	$disk = Get-WmiObject Win32_LogicalDisk -Filter $s | Select-Object Size,FreeSpace
-	$diskUsed = $disk.Size - $disk.FreeSpace
-	$percent = [math]::Round($diskUsed*100 / $disk.Size)
-	Write-Debug "$percent% of $drive is used"
-	return $percent
+    $disk = Get-WmiObject Win32_LogicalDisk -Filter $s | Select-Object Size,FreeSpace
+    $diskUsed = $disk.Size - $disk.FreeSpace
+    $percent = [math]::Round($diskUsed*100 / $disk.Size)
+    Write-Debug "$percent% of $drive is used"
+    return $percent
 }
 
 # Work out the hours to delete ago based on a really complex algorithm!!
 Function Get-HoursAgoToDelete($drive) {
-	[int]$hours = -1000
-	$percentUsed = Get-PercentUsed $drive 
-	$percentUsed = [math]::Round($percentUsed)
+    [int]$hours = -1000
+    $percentUsed = Get-PercentUsed $drive 
+    $percentUsed = [math]::Round($percentUsed)
 
-	if ($percentUsed -lt 51) {$hours = -48}
-	if (($percentUsed -lt 71) -and ($percentUsed -gt 50)) {$hours = -24}
-	if (($percentUsed -lt 81) -and ($percentUsed -gt 70)) {$hours = -12}
-	if (($percentUsed -lt 91) -and ($percentUsed -gt 80)) {$hours = -6}
-	if (($percentUsed -lt 95) -and ($percentUsed -gt 90)) {$hours = -3}
-	if (($percentUsed -lt 100) -and ($percentUsed -gt 94)) {$hours = $failsafeHoursBack}
-	
-	# failsafe to prevent us deleting everything
-	if ($hours -gt $failsafeHoursBack) {
-		$hours = $failsafeHoursBack
-	}
-	# however, if its over a certain percentage, go crazy and delete oldest directory, irrespective of how old
-	if ($percentUsed -gt $percentageAtWhichToPanic) {
-		$hours = $panicHoursBack
-	}
-	Write-Debug "Will attempt to delete oldest folder older than $hours hours"
-	return $hours
+    if ($percentUsed -lt 51) {$hours = -48}
+    if (($percentUsed -lt 71) -and ($percentUsed -gt 50)) {$hours = -24}
+    if (($percentUsed -lt 81) -and ($percentUsed -gt 70)) {$hours = -12}
+    if (($percentUsed -lt 91) -and ($percentUsed -gt 80)) {$hours = -6}
+    if (($percentUsed -lt 95) -and ($percentUsed -gt 90)) {$hours = -3}
+    if (($percentUsed -lt 100) -and ($percentUsed -gt 94)) {$hours = $failsafeHoursBack}
+    
+    # failsafe to prevent us deleting everything
+    if ($hours -gt $failsafeHoursBack) {
+        $hours = $failsafeHoursBack
+    }
+
+    # however, if its over a certain percentage, go crazy and delete oldest directory, irrespective of how old
+    if ($percentUsed -gt $percentageAtWhichToPanic) {
+        $hours = $panicHoursBack
+    }
+
+    Write-Debug "Will attempt to delete oldest folder older than $hours hours"
+    return $hours
 }
 
 
@@ -80,50 +82,55 @@ Function Get-HoursAgoToDelete($drive) {
 # Build Agent Cleanup
 ######################
 Function Run-AgentStrategy($drive) {
-	$hours = Get-HoursAgoToDelete $drive
-	$limit = (Get-Date).AddHours($hours)
-	$folder = $drive + ":\b\"
-	if ($testing) {
-	    Get-ChildItem $folder -Recurse -Depth 0 -ErrorAction SilentlyContinue | Where{$_.Name -Match "^\d{1,4}$"} | ? {$_.LastWriteTime -lt $limit } | Remove-Item -Recurse -WhatIf -Force
-	} else {
- 	    Get-ChildItem $folder -Recurse -Depth 0 -ErrorAction SilentlyContinue | Where{$_.Name -Match "^\d{1,4}$"} | ? {$_.LastWriteTime -lt $limit } | % { cmd.exe /c rmdir /q /s $_.FullName }
-	}
+    $hours = Get-HoursAgoToDelete $drive
+    $limit = (Get-Date).AddHours($hours)
+    $folder = $drive + ":\b\"
+    if ($testing) {
+        Get-ChildItem $folder -Recurse -Depth 0 -ErrorAction SilentlyContinue | Where{$_.Name -Match "^\d{1,4}$"} | ? {$_.LastWriteTime -lt $limit } | Remove-Item -Recurse -WhatIf -Force
+    } else {
+         Get-ChildItem $folder -Recurse -Depth 0 -ErrorAction SilentlyContinue | Where{$_.Name -Match "^\d{1,4}$"} | ? {$_.LastWriteTime -lt $limit } | % { cmd.exe /c rmdir /q /s $_.FullName }
+    }
 }
 
 ################
 # NUGET Cleanup
 ################
 Function Run-NugetStrategy() {
+    $limit = (Get-Date).AddDays($olderThan*-1)
 
-	$limit = (Get-Date).AddDays($olderThan*-1)
+    if ($testing) {
+          $root  = "~\.nuget\packages\"
+        Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force -WhatIf
 
-	if ($testing) {
-	  	$root  = "~\.nuget\packages\"
-		Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force -WhatIf
+          $root  = "~\AppData\Local\NuGet\Cache"    
+        Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force -WhatIf
+    } else {
+          $root  = "~\.nuget\packages\"
+        
+        if (Test-Path $root) {
+            Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force
+        }
 
-	  	$root  = "~\AppData\Local\NuGet\Cache"	
-		Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force -WhatIf
-	} else {
-	  	$root  = "~\.nuget\packages\"
-		Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force
+          $root  = "~\AppData\Local\NuGet\Cache"
 
-	  	$root  = "~\AppData\Local\NuGet\Cache"
-		Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force
-	}
+        if (Test-Path $root) {
+            Get-ChildItem $root | ? { $_.CreationTime -lt $limit } | Remove-Item -Recurse -Force
+        }
+    }
 }
 
 switch ($strategy) {
-	"agent" {
-		Write-Output "Choosing agent strategy, cleans up build agent scratch folders."
-		if ($drive -eq "") {
-			Write-Output "Must supply a drive letter to check."
-		} else {
-		    $DebugPreference = "Continue"
-		    Run-AgentStrategy $drive
-		}
-	}
-	"nuget" {
-		Write-Output "Choosing nuget strategy, cleans up nuget cache folders."
-		Run-NugetStrategy
-	}
+    "agent" {
+        Write-Output "Choosing agent strategy, cleans up build agent scratch folders."
+        if ($drive -eq "") {
+            Write-Output "Must supply a drive letter to check."
+        } else {
+            $DebugPreference = "Continue"
+            Run-AgentStrategy $drive
+        }
+    }
+    "nuget" {
+        Write-Output "Choosing nuget strategy, cleans up nuget cache folders."
+        Run-NugetStrategy
+    }
 }
